Added fogOfWar, back tiles, fore tiles and metadata

+++ SharpDungeon/Game/Tiles/AirTile.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 namespace SharpDungeon.Game.Tiles {
     public class AirTile : TileSingleSide {
 
-        public AirTile(int id) : base(Assets.air, id) {
+        public AirTile(ushort id) : base(Assets.air, id) {
 
         }
